Refactor payment process and add subject ID to payment views
This commit refactors the payment process by adding the subject ID to the payment views. The `PaymentOptionView` and `VisaView` now take in the subject ID as a required parameter. The `CourseDetailsViewBody` widget now passes the subject ID to the `PaymentOptionView` when navigating to it. Additionally, the `VisaViewBody` widget now uses a `StatefulWidget` to set the `WebViewController` and passes the subject ID to the `EnrollmentCubit` when the payment is successful.