Refactor code and fix minor bugs in various files
This commit includes refactoring code and fixing minor bugs in different files across the project. Changes include adding new fields to the SubjectModel class, updating the AppRouter to include new Cubits, updating the QuizViewBody to pass the quiz ID, updating the SearchViewBody to use the ProfileItem widget, and removing the DropDownListCubit from the DropDownListChapterCubit.