Refactor subject list to only show active subjects
The changes in this commit refactor the subject list to only show active subjects in the HomeViewBody widget. A new list is created in the HomeCubit that filters out inactive subjects before emitting the state. The CourseItem widget is now always displayed in the list, regardless of whether the subject is active or not.